// app/api/driver-availability-lookup/driver-availability-table/interface/driver-availability-table.ts
export interface DriverAvailabilityTableEntity {
  is_used?: boolean;
  is_deleted?: boolean;
  is_active: boolean;
  created_at?: string;
  updated_at?: string;
  created_by?: string;
  updated_by?: string;
  host_ip?: string;
  url?: string;
  deleted_at?: string;
  deleted_by?: string;
  tenant_id?: string;
  id: number;
  driver_id: number;
  is_available: boolean;
  current_location_id?: number;
  current_job_id?: number;
  next_free_time?: string;
}